Use 'subdir-objects' option in HAL configure options
If a source file is in a subdirectory, but the 'subdir-objects'
automake option hasn't been enabled. For now, the corresponding
output object file(s) will be placed in the top-level directory.
However, this behaviour will change in future Automake versions,
they will unconditionally cause object files to be placed in the
same subdirectory of the corresponding sources.
So to avoid future incompatibilities, enable the 'subdir-objects'
option in location module configure options

Change-Id: Ie68b822f4cb9eb5c14c623fd6b050ca457886ab1
CRs-Fixed: 2607417